+++ Backend/src/main/java/com/finp/moic/giftCard/application/response/GiftcardBrandServiceResponse.java
@@ -1,0 +1,24 @@
+package com.finp.moic.giftCard.application.response;
+
+import com.querydsl.core.annotations.QueryProjection;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class GiftcardBrandServiceResponse {
+
+    private String mainCategory;
+    private String category;
+
+    public GiftcardBrandServiceResponse() {
+    }
+
+    @QueryProjection
+    @Builder
+    public GiftcardBrandServiceResponse(String mainCategory, String category) {
+        this.mainCategory = mainCategory;
+        this.category = category;
+    }
+}
