--- Backend/src/main/java/com/finp/moic/giftCard/model/dto/request/GiftcardDeleteRequestDTO.java
@@ -1,20 +1,0 @@
-package com.finp.moic.giftCard.model.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class GiftcardDeleteRequestDTO {
-
-    private String imageUrl;
-
-    public GiftcardDeleteRequestDTO() {
-    }
-
-    @Builder
-    public GiftcardDeleteRequestDTO(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-}
