--- Backend/src/main/java/com/finp/moic/giftCard/model/repository/querydsl/GiftcardBrandRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.finp.moic.giftCard.model.repository.querydsl;
-
-import com.finp.moic.giftCard.model.dto.response.GiftcardBrandResponseDTO;
-
-public interface GiftcardBrandRepositoryCustom {
-    GiftcardBrandResponseDTO findByName(String shopName);
-}
