+++ Backend/src/main/java/com/finp/moic/shop/model/repository/ShopRepositoryImpl.java
@@ -2,7 +2,6 @@ package com.finp.moic.shop.model.repository;
 
 import com.finp.moic.card.model.entity.QCardBenefit;
 import com.finp.moic.card.model.entity.QUserCard;
-import com.finp.moic.giftCard.model.entity.QGiftcard;
 import com.finp.moic.shop.model.dto.response.QShopDetailResponseDTO;
 import com.finp.moic.shop.model.dto.response.ShopDetailResponseDTO;
 import com.finp.moic.shop.model.entity.QShop;
