+++ Backend/src/main/java/com/finp/moic/user/model/entity/User.java
@@ -1,6 +1,6 @@
 package com.finp.moic.user.model.entity;
 
-import com.finp.moic.giftCard.model.entity.Giftcard;
+import com.finp.moic.giftCard.domain.Giftcard;
 import com.finp.moic.userBookmark.model.entity.UserBookmark;
 import com.finp.moic.card.model.entity.UserCard;
 import com.finp.moic.util.entity.Base;
